Separating data fetching and UI for Experience; Adding Suspense and Loading for entire site
