+++ app/loading.tsx
@@ -1,0 +1,9 @@
+const Loading = () => {
+  return (
+    <div className="flex h-full w-full items-center justify-center">
+      <div className="border-primary-500 border-t-accent-500 h-10 w-10 animate-spin rounded-full border-4" />
+    </div>
+  );
+};
+
+export default Loading;
